pyproject.toml
[tool.poetry]
name = "factory-boy-demystified"
version = "0.1.0"
description = ""
authors = ["Piotr Kilczuk <piotr@kilczuk.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
Pillow = "^7.0.0"
factory_boy = "^2.12.0"
faker = "^4.0.2"

[tool.poetry.dev-dependencies]
dephell = "^0.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
